requirements.txt
tensorflow>=2.12.0
flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.0.0
seaborn>=0.12.0
opencv-python>=4.8.0
